requirements.txt
tqdm
dateparser
rich
pyzipper
python-dateutil
psd-tools
pillow
pymupdf
